third_party/WebKit/LayoutTests/fast/events/touch/document-create-touch-list-expected.txt

CONSOLE WARNING: line 6: document.createTouchList is deprecated and will be removed in M69, around September 2018. Please use TouchEvent constructor instead. See https://www.chromestatus.com/features/5668612064935936 for more details.
This tests support for the document.createTouchList API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "createTouchList" in document is true
PASS touchList is non-null.
PASS touchList.length is 0
PASS touchList.item(0) is null
PASS touchList.item(1) is null
PASS touchList.item() threw exception TypeError: Failed to execute 'item' on 'TouchList': 1 argument required, but only 0 present..
PASS ts instanceof TouchEvent is true
PASS ts.touches instanceof TouchList is true
PASS ts.touches.length is 2
PASS ts.touches[0] instanceof Touch is true
PASS ts.touches[0].identifier is 12341
PASS ts.touches[0].clientX is 60
PASS ts.touches[1].screenY is 120
PASS ts.ctrlKey is true
PASS An exception was thrown: Failed to execute 'createTouchList' on 'Document': parameter 1 is not of type 'Touch'.
PASS successfullyParsed is true

TEST COMPLETE
PASS successfullyParsed is true

TEST COMPLETE
